Create authenticatedRequestMiddleware as middleware from file-storage-api/accessValidation
