+++ src/routes/file-storage-api/handlers/downloadHandler.ts
@@ -1,8 +1,8 @@
 import { Request, Response, Handler } from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import logger from 'utils/logger';
+import { AuthenticatedRequest } from 'routes/middleware/authenticatedRequestMiddleware';
 import {
-  AuthenticatedRequest,
   hasSufficientDacoAccess,
   hasSufficientProgramMembershipAccess,
 } from '../accessValidations';
