+++ src/routes/file-storage-api/handlers/entitiesIdHandler.ts
@@ -1,7 +1,8 @@
 import { Client } from '@elastic/elasticsearch';
 import { Request, Handler } from 'express';
 import { EsFileCentricDocument } from 'utils/commonTypes/EsFileCentricDocument';
-import { AuthenticatedRequest, hasSufficientProgramMembershipAccess } from '../accessValidations';
+import { AuthenticatedRequest } from 'routes/middleware/authenticatedRequestMiddleware';
+import { hasSufficientProgramMembershipAccess } from '../accessValidations';
 import { getEsFileDocumentByObjectId, toSongEntity } from '../utils';
 
 const createEntitiesIdHandler = ({ esClient }: { esClient: Client }): Handler => {
